Increase modbus_client maximum string storage space (#508)
## [Link to JIRA Ticket](https://flexgen.atlassian.net/browse/ACO-229)

## Description
For really large config files, modbus_client throws the error that
“Total string size up to this point in the config (currently: 65553,
this includes “/components/” per component id) exceeds the maximum of
65535 characters for string storage”. This PR increases the amount of
string storage space available.

## How Has This Been Tested?
Confirmed working against
config/modbus_client/enerc_ess-controller_flexgen-1_modbus_client.json
(32000+ lines!!)

Possible Release Notes: Increase maximum string storage space in
modbus_client

---------

+++ modbus_interface/include/client/client_structs.hpp
@@ -793,7 +793,7 @@ static bool allocate_main_workspace_from_config(const config_loader::Client_Conf
 
     // string storage:
     if (!main_workspace.arena.allocate(main_workspace.string_storage.data, configs_array.total_str_length)) return false;
-    main_workspace.string_storage.allocated = static_cast<u16>(configs_array.total_str_length);
+    main_workspace.string_storage.allocated = static_cast<u32>(configs_array.total_str_length);
     main_workspace.string_storage.current_idx = 0; // for reload purposes (in case this function is called again)
 
     // client workspaces (pointers):
